pyproject.toml
[project]
name = "zepp-export"
version = "0.1.0"
description = "Open-source tools for accessing your Amazfit/Zepp health data"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "requests",
    "python-dotenv",
]

[project.urls]
Repository = "https://github.com/evankp/zepp-export"
Documentation = "https://github.com/evankp/zepp-export/tree/main/docs"
